adding long

+++ cuda-workspace/TestingCudo/grid.h
@@ -36,8 +36,8 @@ typedef struct grid
 	//int yy;
 	uint8_t x;
 	uint8_t y;
-	int count;
-	int iterations;
+	long count;
+	long iterations;
 	} grid;
 
 typedef struct location
